Seerders - Half Done

+++ back-end/mysql/migrations/20200902123038-products.js
@@ -12,7 +12,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       volume: {
